+++ Basil/src/Platform/OpenGL/OpenGLRendererAPI.cpp
@@ -5,10 +5,38 @@
 
 namespace Basil
 {
+	// A callback for logging OpenGL messages
+	void OpenGLMessageCallback(unsigned source, unsigned type, unsigned id, unsigned severity, int length, const char* message, const void* userParam)
+	{
+		switch (severity)
+		{
+			case GL_DEBUG_SEVERITY_HIGH:
+				LOG_FATAL(message);
+				return;
+			case GL_DEBUG_SEVERITY_MEDIUM:
+				LOG_ERROR(message);
+				return;
+			case GL_DEBUG_SEVERITY_LOW:
+				LOG_WARN(message);
+				return;
+			case GL_DEBUG_SEVERITY_NOTIFICATION:
+				LOG_TRACE(message);
+				return;
+		}
+	}
+
 	// Initialise the renderer
 	void OpenGLRendererAPI::init()
 	{
 		PROFILE_FUNCTION();
+
+		// Setup OpenGL logging if running in debug mode
+		#ifdef DEBUG
+				glEnable(GL_DEBUG_OUTPUT);
+				glEnable(GL_DEBUG_OUTPUT_SYNCHRONOUS);
+				glDebugMessageCallback(OpenGLMessageCallback, nullptr);
+				glDebugMessageControl(GL_DONT_CARE, GL_DONT_CARE, GL_DEBUG_SEVERITY_NOTIFICATION, 0, NULL, GL_FALSE);
+		#endif
 
 		// Enable texture blending
 		glEnable(GL_BLEND);