+++ Sandbox/src/SandboxApp.cpp
@@ -12,8 +12,8 @@ class Sandbox : public Basil::Application
 	public:
 		Sandbox()
 		{
-			pushLayer(new ExampleLayer());
-			//pushLayer(new Sandbox2D());
+			//pushLayer(new ExampleLayer());
+			pushLayer(new Sandbox2D());
 		}
 
 		~Sandbox()