+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change log for the Basil Game Engine project.
 ## 2021-04-24
 ### Added
 * The available \_\_FUNCSIG__ defintion is now auto-deduced for wider compiler compatibility.
+* OpenGL messages are now logged in debug mode.
 
 ### Changed
 * Moved ExampleLayer in Sandbox into its own files.